requirements.txt
# Core ML / NLP
numpy==1.26.4
scipy==1.17.1
scikit-learn==1.8.0
torch==2.5.1
transformers==4.42.2
datasets==4.6.1
accelerate==1.12.0
tokenizers==0.19.1
safetensors==0.7.0

# Data / Utils
pandas==2.2.3
pyarrow==23.0.1
joblib==1.4.2
nltk==3.9.1
tqdm==4.67.3
pyyaml==6.0.3

# Experiment Tracking & Pipelines
mlflow==2.17.0
dvc==3.53.0
dvc[s3]
boto3==1.35.36

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
wordcloud==1.9.3

# Serving
Flask==3.0.3
Flask_Cors==5.0.0
lightgbm==4.5.0

-e .


